Differenciation des Texts
git-svn-id: svn+ssh://src.fovea.cc/svn/flobopuyo/trunk@375 053fdb51-1f17-0410-be66-d00040120cd6

+++ PuyoStarter.cpp
@@ -238,7 +238,7 @@ void PuyoGameWidget::cycle()
         }
         if ((cyclesBeforeSpeedIncreases != -1) && (gameSpeed > 0) && (tickCounts % cyclesBeforeSpeedIncreases == 0)) {
             gameSpeed--;
-            cyclesBeforeGameCycleV = (float)cyclesBeforeGameCycle * ((float)gameSpeed / 20.);
+            cyclesBeforeGameCycleV = (int)((float)cyclesBeforeGameCycle * ((float)gameSpeed / 20.));
         }
         
         requestDraw();
